Move Swift version checking logic to Xcode.swift.
