+++ Source/CarthageKitTests/ProjectSpec.swift
@@ -16,23 +16,11 @@ import Tentacle
 class ProjectSpec: QuickSpec {
 	override func spec() {
 		describe("determineBinaryCompatibility") {
-			let currentSwiftVersion = "3.0.2"
-			let testFramework = "Quick.framework"
-			let currentDirectory = URL(fileURLWithPath: FileManager.default.currentDirectoryPath, isDirectory: true)
-			let testFrameworkURL = currentDirectory.appendingPathComponent(testFramework)
-
-			it("should correctly determine current swift verion.") {
-				let result = Project(directoryURL: Bundle(for: type(of: self)).bundleURL).swiftVersion.single()
-				expect(result?.value) == currentSwiftVersion
-			}
-
-			it("should correctly determine a framework's swift version") {
-				let result = Project(directoryURL: Bundle(for: type(of: self)).bundleURL).frameworkSwiftVersion(testFrameworkURL).single()
-
-				expect(result?.value) == currentSwiftVersion
-			}
-
 			it("should pass through framework URLs with the correct Swift version") {
+				let testFramework = "Quick.framework"
+				let currentDirectory = URL(fileURLWithPath: FileManager.default.currentDirectoryPath, isDirectory: true)
+				let testFrameworkURL = currentDirectory.appendingPathComponent(testFramework)
+
 				let result = Project(directoryURL: Bundle(for: type(of: self)).bundleURL).matchingSwiftVersionURL(testFrameworkURL).single()
 
 				expect(result?.value) == testFrameworkURL
@@ -43,7 +31,7 @@ class ProjectSpec: QuickSpec {
 				let result = Project(directoryURL: Bundle(for: type(of: self)).bundleURL).matchingSwiftVersionURL(frameworkURL).single()
 
 				expect(result?.value).to(beNil())
-				expect(result?.error) == .incompatibleFrameworkSwiftVersions(local: currentSwiftVersion, framework: "0.0.0")
+				expect(result?.error) == .incompatibleFrameworkSwiftVersions(local: XcodeSpec.currentSwiftVersion, framework: "0.0.0")
 			}
 		}
